Moved ResourceFinder out of listener package
